figures arranged and figure 4 added
