+++ lineage/Analyze.py
@@ -3,6 +3,7 @@ from concurrent.futures import ProcessPoolExecutor
 import random
 import numpy as np
 from sklearn import metrics
+from sklearn.metrics import balanced_accuracy_score
 from scipy.stats import wasserstein_distance
 import itertools
 
@@ -169,6 +170,7 @@ def Results(tHMMobj, pred_states_by_lineage, LL):
     results_dict["state_proportions_0"] = results_dict["state_proportions"][0]
     results_dict["accuracy_before_switching"] = 100 * np.mean(ravel_pred_states == ravel_true_states)
     results_dict["accuracy_after_switching"] = 100 * np.mean(ravel_switched_pred_states == ravel_true_states)
+    results_dict["balanced_accuracy_score"] = 100 * balanced_accuracy_score(ravel_switched_pred_states, ravel_true_states)
 
     # 4. Calculate the Wasserstein distance
     obs_index = 1
